requirements.txt
aiofiles>=23.2
alembic==1.14
apscheduler>=3.10.1
fastapi==0.115.5
fastparquet
gunicorn
httpx
offsets-db-data @ git+https://github.com/carbonplan/offsets-db-data.git@main #TODO: remove this when the latest version of offsets-db-data is published to pypi
pandas>=1.5.3
psycopg2-binary==2.9.10
pydantic-settings>=2.8
pydantic>=2.10
sqlmodel==0.0.22
ujson
universal_pathlib
uvicorn
fastapi-cache2 @ git+https://github.com/andersy005/fastapi-cache@pydantic-v2-compat # for pydantic v2 compatilibity
watchdog>=3.0
